organized files into webpages folder and components folder for the future
